requirements.txt
Mako==0.9.1
Markdown==2.3.1
MarkupSafe==0.18
PyYAML==3.10
Pygments==1.6
argparse==1.2.1
coverage==3.7.1
mock==1.0.1
nose==1.3.0
pyatom==1.4
simplystatic==0.0.1
wsgiref==0.1.2
